requirements.txt
tqdm
numpy
Pillow
openslide-python
scipy
opencv-python